Rename pixi to rattler_lock_v6
