+++ conda_lockfiles/loaders/__init__.py
@@ -3,18 +3,18 @@ from __future__ import annotations
 from .base import BaseLoader
 from .conda_lock_v1 import CondaLockV1Loader
 from .explicit import ExplicitLoader
-from .pixi import PixiLoader
+from .rattler_lock_v6 import RattlerLockV6Loader
 
 __all__ = [
     "BaseLoader",
     "CondaLockV1Loader",
     "ExplicitLoader",
-    "PixiLoader",
     "LOADERS",
+    "RattlerLockV6Loader",
 ]
 
 LOADERS = (
     CondaLockV1Loader,
     ExplicitLoader,
-    PixiLoader,
+    RattlerLockV6Loader,
 )
