+++ tests/loaders/test_loaders.py
@@ -5,13 +5,13 @@ from conda_lockfiles.loaders import (
     BaseLoader,
     CondaLockV1Loader,
     ExplicitLoader,
-    PixiLoader,
+    RattlerLockV6Loader,
 )
 
 
 def test_loaders() -> None:
     assert LOADERS
     assert BaseLoader not in LOADERS
-    assert PixiLoader in LOADERS
+    assert RattlerLockV6Loader in LOADERS
     assert CondaLockV1Loader in LOADERS
     assert ExplicitLoader in LOADERS
